pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "scram"
version = "0.1.0"
description = "A high-performance terminal-based web crawling agent."
requires-python = ">=3.10"
dependencies = [
    "httpx",
    "curl_cffi",
    "playwright",
    "playwright-stealth",
    "langgraph",
    "langchain",
    "textual",
    "pydantic>=2.0",
    "pandas",
    "google-generativeai",
    "python-dotenv"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
    "mypy"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = [
    ".",
    "src"
]
testpaths = ["tests"]
